# Refactor EIP-712 Typehash Usage to use the new EIP712Constants library for typehashes