I modified the PR to move all ERC-712 hashing related things to a new `ERC712` library. It did end up lowering the codesize a bit, and re-writing the `domainSeparator` logic in assembly also helped with runtime costs, but my personal opinion (after giving much thought to this) is that this is not worth it.

- We don't really have a DX usecase for making sharing these constants and logic:
  - `DOMAIN_TYPEHASH` does not makes sense to share IMO, as it is the domain separator specific to a contract, `Safe`, and not a general constant. Different contracts interacting with Safe should set their own (domains are not shared afterall), and call `domainSeparator` function on the Safe to get the domain
  - We currently don't have a usecase for `SAFE_TX_TYPEHASH`. If you want to compute the Safe transaction hash outside of the `Safe`, you still need to call `domainSeparator`, at which point you might as well call `getTransactionHash`. You _could_ argue that we would be able to share the definition across `Safe` and `CompatibilityFallbackHander`, but (IMO) this is pushing DRY a bit too far (in fact, copying the constant, which won't really ever change, is less code than importing it). It is also a bit of a "compatibility edge case" that I don't think is too important.
  - `SAFE_MSG_TYPEHASH` is specific to the `CompatibilityFallbackHandler`, and the `ExtensibleFallbackHAndler` also implements it for compatibility purposes, so I would argue similar reasoning to the above point.
- Moving ERC-712 functions to a separate library can be useful, but it a more substantial code change that I am not sure is worth it at the moment.

Anyway, I pushed some changes so others can take a look at my experimentation, but I will close this PR for now.

I will make a follow up re-implementing the `domainSeparator` computation in assembly as I think that it is generally worth it (especially because it is on the "critical path" for Safe transaction execution, and it is fairly simple code).

<details><summary>Old PR description</summary>

This pull request introduces a refactoring across multiple contracts to centralise EIP-712 type hash constants into a new `EIP712Constants` library. This change improves maintainability and reduces redundancy by replacing hardcoded constants with references to the centralised library and improves the developer experience by exporting these constants.

Implements https://github.com/safe-global/safe-smart-account/issues/974

**Benchmark before:**
```
  Safe
           Used 166287n gas for >Setup Safe with 1 owner(s) and fallback handler<
           Used 166287n gas for >Setup Safe with 1 owner(s) and fallback handler<
           Used 189722n gas for >Setup Safe with 2 owner(s) and fallback handler<
           Used 213145n gas for >Setup Safe with 3 owner(s) and fallback handler<
           Used 260015n gas for >Setup Safe with 5 owner(s) and fallback handler<
    ✔ creation (541ms)

  Proxy - creation
           Used 105280n gas for >creation<
    ✔ with an EOA (76ms)
           Used 136002n gas for >creation<
    ✔ with a single owner Safe
           Used 141907n gas for >creation<
    ✔ with a single owner and guard Safe
           Used 143051n gas for >creation<
    ✔ with a 2 out of 2 Safe
           Used 150099n gas for >creation<
    ✔ with a 3 out of 3 Safe
           Used 150099n gas for >creation<
    ✔ with a 3 out of 5 Safe

  Proxy - chain specific creation
           Used 105466n gas for >chain specific creation<
    ✔ with an EOA (72ms)
           Used 136188n gas for >chain specific creation<
    ✔ with a single owner Safe
           Used 142093n gas for >chain specific creation<
    ✔ with a single owner and guard Safe
           Used 143225n gas for >chain specific creation<
    ✔ with a 2 out of 2 Safe
           Used 150273n gas for >chain specific creation<
    ✔ with a 3 out of 3 Safe
           Used 150285n gas for >chain specific creation<
    ✔ with a 3 out of 5 Safe

  ERC20 - transfer
           Used 51800n gas for >transfer<
    ✔ with an EOA (74ms)
           Used 82498n gas for >transfer<
    ✔ with a single owner Safe
           Used 88379n gas for >transfer<
    ✔ with a single owner and guard Safe
           Used 89546n gas for >transfer<
    ✔ with a 2 out of 2 Safe
           Used 96595n gas for >transfer<
    ✔ with a 3 out of 3 Safe
           Used 96595n gas for >transfer<
    ✔ with a 3 out of 5 Safe

  Ether - transfer
           Used 21000n gas for >transfer<
    ✔ with an EOA (69ms)
           Used 58244n gas for >transfer<
    ✔ with a single owner Safe
           Used 64088n gas for >transfer<
    ✔ with a single owner and guard Safe
           Used 65292n gas for >transfer<
    ✔ with a 2 out of 2 Safe
           Used 72328n gas for >transfer<
    ✔ with a 3 out of 3 Safe
           Used 72340n gas for >transfer<
    ✔ with a 3 out of 5 Safe

  MultiSend - multiple ERC20 transfers
    - with an EOA
           Used 92709n gas for >multiple ERC20 transfers<
    ✔ with a single owner Safe
           Used 98765n gas for >multiple ERC20 transfers<
    ✔ with a single owner and guard Safe
           Used 99760n gas for >multiple ERC20 transfers<
    ✔ with a 2 out of 2 Safe
           Used 106799n gas for >multiple ERC20 transfers<
    ✔ with a 3 out of 3 Safe
           Used 106799n gas for >multiple ERC20 transfers<
    ✔ with a 3 out of 5 Safe

  ERC1155 - transfer
           Used 57234n gas for >transfer<
    ✔ with an EOA (75ms)
           Used 87981n gas for >transfer<
    ✔ with a single owner Safe
           Used 93934n gas for >transfer<
    ✔ with a single owner and guard Safe
           Used 95053n gas for >transfer<
    ✔ with a 2 out of 2 Safe
           Used 102104n gas for >transfer<
    ✔ with a 3 out of 3 Safe
           Used 102092n gas for >transfer<
    ✔ with a 3 out of 5 Safe

```

**Benchmark after:**
```
  Safe
           Used 166287n gas for >Setup Safe with 1 owner(s) and fallback handler<
           Used 166287n gas for >Setup Safe with 1 owner(s) and fallback handler<
           Used 189722n gas for >Setup Safe with 2 owner(s) and fallback handler<
           Used 213145n gas for >Setup Safe with 3 owner(s) and fallback handler<
           Used 260015n gas for >Setup Safe with 5 owner(s) and fallback handler<
    ✔ creation (525ms)

  Proxy - creation
           Used 105280n gas for >creation<
    ✔ with an EOA (76ms)
           Used 136021n gas for >creation<
    ✔ with a single owner Safe
           Used 141926n gas for >creation<
    ✔ with a single owner and guard Safe
           Used 143070n gas for >creation<
    ✔ with a 2 out of 2 Safe
           Used 150106n gas for >creation<
    ✔ with a 3 out of 3 Safe
           Used 150118n gas for >creation<
    ✔ with a 3 out of 5 Safe

  Proxy - chain specific creation
           Used 105466n gas for >chain specific creation<
    ✔ with an EOA (70ms)
           Used 136207n gas for >chain specific creation<
    ✔ with a single owner Safe
           Used 142112n gas for >chain specific creation<
    ✔ with a single owner and guard Safe
           Used 143256n gas for >chain specific creation<
    ✔ with a 2 out of 2 Safe
           Used 150292n gas for >chain specific creation<
    ✔ with a 3 out of 3 Safe
           Used 150292n gas for >chain specific creation<
    ✔ with a 3 out of 5 Safe

  ERC20 - transfer
           Used 51800n gas for >transfer<
    ✔ with an EOA (71ms)
           Used 82517n gas for >transfer<
    ✔ with a single owner Safe
           Used 88386n gas for >transfer<
    ✔ with a single owner and guard Safe
           Used 89541n gas for >transfer<
    ✔ with a 2 out of 2 Safe
           Used 96590n gas for >transfer<
    ✔ with a 3 out of 3 Safe
           Used 96602n gas for >transfer<
    ✔ with a 3 out of 5 Safe

  Ether - transfer
           Used 21000n gas for >transfer<
    ✔ with an EOA (66ms)
           Used 58263n gas for >transfer<
    ✔ with a single owner Safe
           Used 64107n gas for >transfer<
    ✔ with a single owner and guard Safe
           Used 65311n gas for >transfer<
    ✔ with a 2 out of 2 Safe
           Used 72359n gas for >transfer<
    ✔ with a 3 out of 3 Safe
           Used 72323n gas for >transfer<
    ✔ with a 3 out of 5 Safe

  MultiSend - multiple ERC20 transfers
    - with an EOA
           Used 92728n gas for >multiple ERC20 transfers<
    ✔ with a single owner Safe
           Used 98784n gas for >multiple ERC20 transfers<
    ✔ with a single owner and guard Safe
           Used 99779n gas for >multiple ERC20 transfers<
    ✔ with a 2 out of 2 Safe
           Used 106830n gas for >multiple ERC20 transfers<
    ✔ with a 3 out of 3 Safe
           Used 106830n gas for >multiple ERC20 transfers<
    ✔ with a 3 out of 5 Safe

  ERC1155 - transfer
           Used 57234n gas for >transfer<
    ✔ with an EOA (71ms)
           Used 88012n gas for >transfer<
    ✔ with a single owner Safe
           Used 93953n gas for >transfer<
    ✔ with a single owner and guard Safe
           Used 95072n gas for >transfer<
    ✔ with a 2 out of 2 Safe
           Used 102123n gas for >transfer<
    ✔ with a 3 out of 3 Safe
           Used 102111n gas for >transfer<
    ✔ with a 3 out of 5 Safe
```

</details>