@mmv08 2025-06-25
> In general I like the change, I would just propose that we make a new branch for proposing these changes - I don't want to use main since we are getting ready for 1.5.0 release.

We can also create a release branch for 1.5.0. In a sense the change is already proposed in a new branch :D